requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.10

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# External Services
cloudinary==1.36.0
redis==5.0.1
httpx==0.25.2

# Image Processing
Pillow==11.3.0

# Data Validation
pydantic==2.11.5
pydantic-settings==2.1.0

# Utilities
python-dotenv==1.0.0